+++ CHANGELOG.md
@@ -104,6 +104,9 @@ Adding a new version? You'll need three changes:
   - `FullPathRewrite` [#5855](https://github.com/Kong/kubernetes-ingress-controller/pull/5855)
 - DB mode now supports Event reporting for resources that failed to apply.
   [#5785](https://github.com/Kong/kubernetes-ingress-controller/pull/5785)
+- The Konnect license agent now attempts to update more frequently if its
+  license is expired.
+  [#5936](https://github.com/Kong/kubernetes-ingress-controller/pull/5936)
 
 ### Fixed